# schemas/api/openapi.yaml
openapi: 3.1.0
info:
  title: NextSteamGame API
  version: 1.0.0
  description: Versioned public contract for deterministic game recommendations.
servers:
  - url: http://localhost:8080
paths:
  /healthz:
    get:
      operationId: health
      responses:
        "200":
          description: Process is alive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /readyz:
    get:
      operationId: readiness
      responses:
        "200":
          description: Service loaded and validated one immutable build.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"
        "503":
          $ref: "#/components/responses/Error"
  /v1/config:
    get:
      operationId: getConfig
      responses:
        "200":
          description: Public product and request limits.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfigResponse"
  /v1/games/search:
    get:
      operationId: searchGames
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 100
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Title matches in deterministic display order.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameSearchResponse"
        "400":
          $ref: "#/components/responses/Error"
  /v1/games/{appid}:
    get:
      operationId: getGame
      parameters:
        - $ref: "#/components/parameters/AppId"
      responses:
        "200":
          description: Public game metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          $ref: "#/components/responses/Error"
  /v1/recommendations:
    post:
      operationId: recommendGames
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationRequest"
      responses:
        "200":
          description: Deterministic ranking for normalized intent and active build.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationResponse"
        "400":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
components:
  parameters:
    AppId:
      name: appid
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/GameId"
  responses:
    Error:
      description: Typed request or service error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    GameId:
      type: integer
      format: uint32
      minimum: 1
    BuildId:
      type: string
      minLength: 1
      maxLength: 128
    Lane:
      type: string
      enum: [mechanics, narrative, vibe, structure_loop]
    LaneWeights:
      type: object
      additionalProperties: false
      required: [mechanics, narrative, vibe, structure_loop]
      properties:
        mechanics:
          type: number
          minimum: 0
          maximum: 2
        narrative:
          type: number
          minimum: 0
          maximum: 2
        vibe:
          type: number
          minimum: 0
          maximum: 2
        structure_loop:
          type: number
          minimum: 0
          maximum: 2
    Seed:
      type: object
      additionalProperties: false
      required: [appid, weight]
      properties:
        appid:
          $ref: "#/components/schemas/GameId"
        weight:
          type: number
          exclusiveMinimum: 0
          maximum: 2
    IntentInput:
      type: object
      additionalProperties: false
      required: [lane_weights, include, exclude, text]
      properties:
        lane_weights:
          $ref: "#/components/schemas/LaneWeights"
        include:
          type: array
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 80
        exclude:
          type: array
          maxItems: 20
          uniqueItems: true
          items:
            type: string
            minLength: 1
            maxLength: 80
        text:
          type: [string, "null"]
          maxLength: 500
    RecommendationRequest:
      type: object
      additionalProperties: false
      required: [seeds, intent, limit]
      properties:
        seeds:
          type: array
          minItems: 1
          maxItems: 8
          items:
            $ref: "#/components/schemas/Seed"
        intent:
          $ref: "#/components/schemas/IntentInput"
        limit:
          type: integer
          minimum: 1
          maximum: 100
    NormalizedIntent:
      allOf:
        - $ref: "#/components/schemas/IntentInput"
        - type: object
          required: [seeds]
          properties:
            seeds:
              type: array
              items:
                $ref: "#/components/schemas/Seed"
    Game:
      type: object
      additionalProperties: false
      required: [appid, name, short_description, steam_url]
      properties:
        appid:
          $ref: "#/components/schemas/GameId"
        name:
          type: string
        short_description:
          type: string
        release_year:
          type: [integer, "null"]
          minimum: 1970
          maximum: 2100
        header_image_url:
          type: [string, "null"]
          format: uri
        steam_url:
          type: string
          format: uri
    GameSearchResponse:
      type: object
      additionalProperties: false
      required: [games]
      properties:
        games:
          type: array
          items:
            $ref: "#/components/schemas/Game"
    Evidence:
      type: object
      additionalProperties: false
      required: [id, appid, source, excerpt]
      properties:
        id:
          type: string
        appid:
          $ref: "#/components/schemas/GameId"
        source:
          type: string
        excerpt:
          type: string
          maxLength: 500
    ScoreContribution:
      type: object
      additionalProperties: false
      required: [key, label, value, evidence_ids]
      properties:
        key:
          type: string
        label:
          type: string
        lane:
          oneOf:
            - $ref: "#/components/schemas/Lane"
            - type: "null"
        value:
          type: number
        evidence_ids:
          type: array
          items:
            type: string
    Recommendation:
      type: object
      additionalProperties: false
      required: [rank, game, score, contributions, matched_concepts, explanation, evidence]
      properties:
        rank:
          type: integer
          minimum: 1
        game:
          $ref: "#/components/schemas/Game"
        score:
          type: number
        contributions:
          type: array
          items:
            $ref: "#/components/schemas/ScoreContribution"
        matched_concepts:
          type: array
          items:
            type: string
        explanation:
          type: string
        evidence:
          type: array
          items:
            $ref: "#/components/schemas/Evidence"
    RetrievalState:
      type: object
      additionalProperties: false
      required: [mode, reduced_confidence]
      properties:
        mode:
          type: string
          enum: [candidate_graph, dynamic_vector, sparse_fallback]
        reduced_confidence:
          type: boolean
        reason:
          type: [string, "null"]
    RecommendationResponse:
      type: object
      additionalProperties: false
      required: [build_id, normalized_intent, retrieval, results]
      properties:
        build_id:
          $ref: "#/components/schemas/BuildId"
        normalized_intent:
          $ref: "#/components/schemas/NormalizedIntent"
        retrieval:
          $ref: "#/components/schemas/RetrievalState"
        results:
          type: array
          items:
            $ref: "#/components/schemas/Recommendation"
    HealthResponse:
      type: object
      additionalProperties: false
      required: [status, build_id]
      properties:
        status:
          type: string
          const: ok
        build_id:
          $ref: "#/components/schemas/BuildId"
    ReadyResponse:
      type: object
      additionalProperties: false
      required: [status, build_id]
      properties:
        status:
          type: string
          const: ready
        build_id:
          $ref: "#/components/schemas/BuildId"
    ConfigResponse:
      type: object
      additionalProperties: false
      required: [build_id, max_results, max_body_bytes, lanes]
      properties:
        build_id:
          $ref: "#/components/schemas/BuildId"
        max_results:
          type: integer
        max_body_bytes:
          type: integer
        lanes:
          type: array
          items:
            $ref: "#/components/schemas/Lane"
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, request_id]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
        request_id:
          type: string